Support for building with fips support
Passes the FIPS=1 value into cmake to build the library in FIPS mode.
https://boringssl.googlesource.com/boringssl/+/fed35d32245ee4563691d21f55c12b4f8dac840a/crypto/fipsmodule/FIPS.md

Since the top level cmake file does not accept the fips variable we need
to point the build towards the cmake file in the src directory and
update the search paths for the crypto and ssl libraries respectively.